# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Proof Node API",
        "description": "HTTP API for CL to request execution proof generation/verification of different EL/zkVM combinations. Proofs are generated asynchronously; Server-Sent Events notify when a proof request completes or fails.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Default local server"
        }
    ],
    "paths": {
        "/v1/execution_proof_requests": {
            "post": {
                "operationId": "postExecutionProofRequests",
                "summary": "Submit an execution proof request",
                "description": "Accepts an SSZ-encoded ProofRequestBody (payload, chain config, and proof types) and enqueues proof generation. Returns the tree-hash root computed from the submitted payload.",
                "requestBody": {
                    "required": true,
                    "description": "SSZ-encoded ProofRequestBody containing the NewPayloadRequest (transparent union across fork variants), the chain config, and the proof types to generate.",
                    "content": {
                        "application/octet-stream": {
                            "schema": {
                                "$ref": "#/components/schemas/ProofRequestBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proof request accepted.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProofRequestResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid SSZ body, empty or missing proof_types, or unconfigured proof type.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "examples": {
                                    "invalid_ssz": {
                                        "summary": "Invalid SSZ body",
                                        "value": {
                                            "code": 400,
                                            "message": "invalid SSZ body: OffsetOutOfBounds(0)"
                                        }
                                    },
                                    "unconfigured_proof_type": {
                                        "summary": "Unconfigured proof type",
                                        "value": {
                                            "code": 400,
                                            "message": "no zkVM configured for proof type 'reth-sp1'"
                                        }
                                    },
                                    "empty_proof_types": {
                                        "summary": "Empty proof_types",
                                        "value": {
                                            "code": 400,
                                            "message": "proof_types must contain at least one valid value"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal queue failure.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "code": 500,
                                    "message": "failed to enqueue proof: channel closed"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "operationId": "getExecutionProofRequests",
                "summary": "Subscribe to execution proof events via SSE",
                "description": "Opens a Server-Sent Events stream for proof completion and failure notifications.\n\nThe server sends SSE comment lines (`:` prefix, per SSE spec) every 15 seconds as keep-alive. These are not events and should be ignored by conforming SSE clients.\n\nWhen `new_payload_request_root` is provided, the server first emits cached `proof_complete` events for cached proofs that completed before the subscription started (catch-up), then streams live events filtered to that root.\n\nCompleted proofs are held in a bounded LRU cache. Clients should download proofs promptly after receiving a `proof_complete` event, as proofs may be evicted under memory pressure.\n\n**Event types:**\n- `proof_complete` - Data payload: ProofCompleteEvent (JSON)\n- `proof_failure` - Data payload: ProofFailureEvent (JSON)",
                "parameters": [
                    {
                        "name": "new_payload_request_root",
                        "in": "query",
                        "required": false,
                        "description": "Filter events to only this root. When omitted, all events are streamed.",
                        "schema": {
                            "$ref": "#/components/schemas/Hash256"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "SSE event stream opened. Events are delivered as `text/event-stream`.",
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "description": "SSE stream. Each message has an `event` field (event type name) and a `data` field (JSON payload). See the SSE event schemas for the payload format of each event type.",
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/execution_proofs/{new_payload_request_root}/{proof_type}": {
            "get": {
                "operationId": "getExecutionProofs",
                "summary": "Download a completed execution proof",
                "description": "Returns the raw proof bytes for a completed proof identified by its root and proof type.",
                "parameters": [
                    {
                        "name": "new_payload_request_root",
                        "in": "path",
                        "required": true,
                        "description": "The root from the original proof request response.",
                        "schema": {
                            "$ref": "#/components/schemas/Hash256"
                        }
                    },
                    {
                        "name": "proof_type",
                        "in": "path",
                        "required": true,
                        "description": "The proof type to download.",
                        "schema": {
                            "$ref": "#/components/schemas/ProofType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Raw proof bytes.",
                        "content": {
                            "application/octet-stream": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proof not yet available or does not exist.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "code": 404,
                                    "message": "proof not found for root 0xabcd...1234 and type reth-sp1"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Hash256": {
                "type": "string",
                "pattern": "^0x[0-9a-f]{64}$",
                "minLength": 66,
                "maxLength": 66,
                "description": "A 32-byte hash represented as a 0x-prefixed lowercase hex string (66 characters total). The 0x prefix is required in both requests and responses.",
                "examples": [
                    "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
                ]
            },
            "ProofType": {
                "type": "string",
                "enum": [
                    "ethrex-risc0",
                    "ethrex-sp1",
                    "ethrex-zisk",
                    "reth-openvm",
                    "reth-risc0",
                    "reth-sp1",
                    "reth-zisk"
                ],
                "description": "Case-sensitive identifier for an EL/zkVM combination."
            },
...
            "ProofRequestResponse": {
                "type": "object",
                "required": [
                    "new_payload_request_root"
                ],
                "properties": {
                    "new_payload_request_root": {
                        "$ref": "#/components/schemas/Hash256"
                    }
                },
                "description": "Response for a successful proof request submission."
            },
...
            "ErrorResponse": {
                "type": "object",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "description": "HTTP status code."
                    },
                    "message": {
                        "type": "string",
                        "description": "Human-readable error message."
                    }
                },
                "description": "Standard error response returned by all endpoints on failure."
            },
...
            "NewPayloadRequest": {
                "type": "object",
                "description": "Consensus-layer new-payload request, an SSZ transparent union across fork variants (Bellatrix, Capella, Deneb, Electra/Fulu, Gloas). See the consensus-specs NewPayloadRequest for the per-variant field layout.",
                "properties": {
                    "execution_payload": {
                        "type": "object",
                        "description": "The fork-dependent ExecutionPayload."
                    }
                }
            },
            "ProtocolFork": {
                "type": "integer",
                "format": "uint64",
                "description": "SSZ enum discriminant selecting the active execution-layer fork, in fork order"
            },
            "BlobSchedule": {
                "type": "object",
                "required": [
                    "target",
                    "max",
                    "base_fee_update_fraction"
                ],
                "properties": {
                    "target": {
                        "type": "integer",
                        "format": "uint64",
                        "description": "Target blobs per block."
                    },
                    "max": {
                        "type": "integer",
                        "format": "uint64",
                        "description": "Maximum blobs per block."
                    },
                    "base_fee_update_fraction": {
                        "type": "integer",
                        "format": "uint64",
                        "description": "Blob base fee update fraction."
                    }
                },
                "description": "EIP-7840 blob parameters for the active fork. The consensus layer provides max, which the server validates against its execution-layer genesis. target and base_fee_update_fraction may be 0 while the consensus config does not expose them, in which case the server fills them from the execution-layer genesis."
            },
            "ForkActivation": {
                "type": "object",
                "description": "Activation point of the fork. At least one of block_number or timestamp is set, each SSZ-encoded as a list of zero or one element.",
                "properties": {
                    "block_number": {
                        "type": "array",
                        "maxItems": 1,
                        "items": {
                            "type": "integer",
                            "format": "uint64"
                        },
                        "description": "Activation block number, for pre-merge forks."
                    },
                    "timestamp": {
                        "type": "array",
                        "maxItems": 1,
                        "items": {
                            "type": "integer",
                            "format": "uint64"
                        },
                        "description": "Activation timestamp, for post-merge forks."
                    }
                }
            },
            "ForkConfig": {
                "type": "object",
                "required": [
                    "fork",
                    "activation",
                    "blob_schedule"
                ],
                "properties": {
                    "fork": {
                        "$ref": "#/components/schemas/ProtocolFork"
                    },
                    "activation": {
                        "$ref": "#/components/schemas/ForkActivation"
                    },
                    "blob_schedule": {
                        "type": "array",
                        "maxItems": 1,
                        "items": {
                            "$ref": "#/components/schemas/BlobSchedule"
                        },
                        "description": "Zero or one blob schedule. Present for blob-bearing forks (Cancun onward), empty otherwise."
                    }
                },
                "description": "The protocol fork active for the block, with its activation point and blob schedule."
            },
            "ChainConfig": {
                "type": "object",
                "required": [
                    "chain_id",
                    "active_fork"
                ],
                "properties": {
                    "chain_id": {
                        "type": "integer",
                        "format": "uint64",
                        "description": "EIP-155 chain id."
                    },
                    "active_fork": {
                        "$ref": "#/components/schemas/ForkConfig"
                    }
                },
                "description": "The chain configuration the proof is verified against. It carries chain_id and the resolved active fork."
            },
            "ProofRequestBody": {
                "type": "object",
                "required": [
                    "new_payload_request",
                    "chain_config",
                    "proof_types"
                ],
                "properties": {
                    "new_payload_request": {
                        "$ref": "#/components/schemas/NewPayloadRequest"
                    },
                    "chain_config": {
                        "$ref": "#/components/schemas/ChainConfig"
                    },
                    "proof_types": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProofType"
                        },
                        "description": "Proof types to generate for this payload. Must be non-empty with no duplicates."
                    }
                },
                "description": "Logical field layout of the body for POST /v1/execution_proof_requests. The wire format is SSZ (application/octet-stream), not JSON."
            },
...
        }
    }
}